+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/tasks/lifecycle/GenerateTask.kt
@@ -7,7 +7,9 @@ import org.gradle.api.plugins.BasePlugin.ASSEMBLE_TASK_NAME
 import org.gradle.api.plugins.BasePlugin.BUILD_GROUP
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.kotlin.dsl.register
+import org.gradle.work.DisableCachingByDefault
 
+@DisableCachingByDefault
 public open class GenerateTask : DefaultTask(), HubdleTask {
 
     init {